+++ backend/carbook/src/main/java/softeer/carbook/domain/user/controller/UserController.java
@@ -17,7 +17,6 @@ import javax.validation.Valid;
 
 @RestController
 public class UserController {
-    private static final Logger logger = LoggerFactory.getLogger(UserController.class);
     private final UserService userService;
 
     @Autowired
@@ -73,53 +72,5 @@ public class UserController {
                 httpServletRequest);
         return Message.make200Response(resultMsg.getMessage());
     }
-
-    // exception handling
-
-    // Valid 어노테이션 예외 처리
-    @ExceptionHandler(BindException.class)
-    public ResponseEntity<Message> processValidationError(BindException ex) {
-        BindingResult bindingResult = ex.getBindingResult();
-        String errorMsg = bindingResult.getFieldError().getDefaultMessage();
-        logger.warn(errorMsg);
-        return Message.make400Response(errorMsg);
-    }
-
-
-    // 회원가입 시 이메일 중복 처리
-    @ExceptionHandler(SignupEmailDuplicateException.class)
-    public ResponseEntity<Message> signupEmailDuplicateException(SignupEmailDuplicateException emailDE){
-        logger.debug(emailDE.getMessage());
-        return Message.make400Response(emailDE.getMessage());
-    }
-
-    // 회원가입, 닉네임 변경 시 닉네임 중복 처리
-    @ExceptionHandler(NicknameDuplicateException.class)
-    public ResponseEntity<Message> nicknameDuplicateException(NicknameDuplicateException nicknameDE){
-        logger.debug(nicknameDE.getMessage());
-        return Message.make400Response(nicknameDE.getMessage());
-    }
-
-    // 로그인 시 등록된 이메일이 없는 경우 처리
-    @ExceptionHandler(LoginEmailNotExistException.class)
-    public ResponseEntity<Message> loginEmailNotExistException(LoginEmailNotExistException emailNE){
-        logger.debug(emailNE.getMessage());
-        return Message.make400Response(emailNE.getMessage());
-    }
-
-    // 닉네임이 데이터베이스에 존재하지 않는 경우 처리
-    @ExceptionHandler(NicknameNotExistException.class)
-    public ResponseEntity<Message> nicknameNotExistException(NicknameNotExistException nicknameNE){
-        logger.debug(nicknameNE.getMessage());
-        return Message.make400Response(nicknameNE.getMessage());
-    }
-
-    // 로그인 상태가 아닌 경우 처리
-    @ExceptionHandler(NotLoginStatementException.class)
-    public ResponseEntity<Message> notLoginStatementException(NotLoginStatementException notLoginE){
-        logger.debug(notLoginE.getMessage());
-        return Message.make400Response(notLoginE.getMessage());
-    }
-
 
 }
